+++ R/colorMaps.R
@@ -187,3 +187,30 @@ colorMap4 <- function(
   P
 }
 
+#' @rdname colorMaps
+#' @export
+colorMap5 <- function(
+    Z, bkgcolor = "#15191e", nancolor = "#000000", 
+    reverse = c(FALSE, FALSE, FALSE),
+    nthreads = 1L
+){
+  stopifnot(isComplex(Z))
+  stopifnot(isString(bkgcolor))
+  stopifnot(isString(nancolor))
+  stopifnot(isBooleanTriplet(reverse))
+  nthreads <- as.integer(nthreads)
+  stopifnot(nthreads >= 1L)
+  storage.mode(Z) <- "complex"
+  ismatrix <- is.matrix(Z)
+  if(!ismatrix){
+    Z <- cbind(Z)
+  }
+  P <- ColorMap5(
+    Z, bkgcolor, nancolor, reverse[1], reverse[2], reverse[3], nthreads
+  )
+  if(!ismatrix){
+    P <- c(P)
+  }
+  P
+}
+